Added Scene Doors + Interactable Ray + Levels
Scene door that are an inherited of Interactable can now be use by a raycast on the player camera. Leaving it open ended in case we want to add; buttons, items, etc.

I also made every level a "Level" class, so for all new levels, lets make sure to make an inherited scene of "Level." This should open a few customizable options and set a standard going forward for level design.